+++ flutter_inappwebview_android/android/src/main/java/com/pichillilorenzo/flutter_inappwebview_android/webview/WebViewChannelDelegateMethods.java
@@ -85,5 +85,7 @@ public enum WebViewChannelDelegateMethods {
   isInFullscreen,
   clearFormData,
   hideInputMethod,
-  showInputMethod
+  showInputMethod,
+  saveState,
+  restoreState,
 }