+++ flutter_inappwebview_ios/ios/Classes/InAppWebView/WebViewChannelDelegate.swift
@@ -691,6 +691,22 @@ public class WebViewChannelDelegate: ChannelDelegate {
             }
             result(true)
             break
+        case .saveState:
+            if let webView = webView, #available(iOS 15.0, *) {
+                result(webView.saveState())
+            } else {
+                result(nil)
+            }
+            break
+        case .restoreState:
+            if let webView = webView, #available(iOS 15.0, *) {
+                let state = arguments!["state"] as! FlutterStandardTypedData
+                webView.restoreState(state: state.data)
+                result(true)
+            } else {
+                result(false)
+            }
+            break
         }
     }