+++ flutter_inappwebview_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0-beta.3
+
+- Updated flutter_inappwebview_platform_interface version to ^1.4.0-beta.3
+- Implemented `saveState`, `restoreState` InAppWebViewController methods
+
 ## 1.2.0-beta.2
 
 - Updated flutter_inappwebview_platform_interface version to ^1.4.0-beta.2